feat : 책 추가 기능에 분류명 추가
issue : #80

+++ src/main/java/keeper/project/homepage/dto/library/BookDto.java
@@ -17,7 +17,7 @@ public class BookDto {
   private String title;
   private String author;
   private String information;
-  private Integer department;
+  private Long department;
   private Long quantity;
   private Long thumbnailId;
